Silence numerous warnings in nss/cmd, most of which could be real problems. on Hacks branch only.
--HG--
branch : NSS_PERFORMANCE_HACKS_BRANCH

+++ security/nss/cmd/strsclnt/strsclnt.c
@@ -400,7 +400,7 @@ thread_wrapper(void * arg)
 
     /* Handle cleanup of thread here. */
     PRINTF("strsclnt: Thread in slot %d returned %d\n", 
-	   slot - threads, slot->rv);
+	   (int)(slot - threads), slot->rv);
 
     PR_Lock(threadLock);
     slot->running = rs_idle;
